requirements.txt
yfinance==0.2.18
beautifulsoup4==4.12.2
requests==2.28.2
pandas==2.0.0
google-cloud-bigquery==3.7.0
google-cloud-storage==2.9.0
python-dotenv==1.0.0
db-dtypes==1.1.1
gunicorn==20.1.0
Flask==2.3.1
Flask-Cors==3.0.10
requests-ratelimiter==0.4.0
pyrate-limiter==2.10.0
pytz==2023.3